airflow/requirements.txt
apache-airflow
azure-storage-blob
hubspot-api-client
llama-index
pymilvus
